show_logs better flag support
not yet implemented in the depency log functions show showing a logfile
starting with dashes is still not working but the base layer is there

+++ show_log.sh
@@ -12,6 +12,7 @@ source lib/lib.sh
 arg_logpath=""
 arg_follow=""
 arg_print=0
+arg_no_args=0
 
 for arg in "$@"
 do
@@ -25,30 +26,39 @@ do
 		echo "  --tem       ignore logname when used as TeeworldsEconMod input"
 		echo "  logname     directory in ./logs/logname for local log types"
 		exit 0
-	elif [ "$arg" == "--filepath" ]
+	elif [ "$arg" == "--" ]
 	then
-		arg_print=1
-	elif [ "$arg" == "--tem" ]
+		arg_no_args=1
+	elif [ "${arg::1}" == "-" ] && [ "$arg_no_args" == "0" ]
 	then
-		arg_follow=-f
-		uuid="$(generate_uuid)"
-		current_file="$(get_latest_logfile)"
-		show_latest_log "$arg_follow" "$arg_logpath" --id="$uuid" &
-		while true
-		do
-			if [ "$current_file" != "$(get_latest_logfile)" ]
-			then
-				pkill -f "id=$uuid"
-				pkill -f "tail.*$current_file"
-				current_file="$(get_latest_logfile)"
-				show_latest_log "$arg_follow" "$arg_logpath" --id="$uuid" &
-			fi
-			sleep 1
-		done
-		break
-	elif [ "$arg" == "-f" ]
-	then
-		arg_follow=-f
+		if [ "$arg" == "--filepath" ]
+		then
+			arg_print=1
+		elif [ "$arg" == "--tem" ]
+		then
+			arg_follow=-f
+			uuid="$(generate_uuid)"
+			current_file="$(get_latest_logfile)"
+			show_latest_log "$arg_follow" "$arg_logpath" --id="$uuid" &
+			while true
+			do
+				if [ "$current_file" != "$(get_latest_logfile)" ]
+				then
+					pkill -f "id=$uuid"
+					pkill -f "tail.*$current_file"
+					current_file="$(get_latest_logfile)"
+					show_latest_log "$arg_follow" "$arg_logpath" --id="$uuid" &
+				fi
+				sleep 1
+			done
+			break
+		elif [ "$arg" == "-f" ]
+		then
+			arg_follow=-f
+		else
+			err "Error: unkown flag '$arg' see '--help'"
+			exit 1
+		fi
 	elif [ "$logpath" == "" ]
 	then
 		arg_logpath="$arg"
